moved out the ai stuff
